+++ src/ModalManager.js
@@ -1,4 +1,3 @@
-import { get, upperFirst } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
@@ -7,6 +6,12 @@ import {
   injectIntl,
 } from 'react-intl';
 import {
+  get,
+  lowerFirst,
+  upperFirst,
+} from 'lodash';
+
+import {
   ConfirmationModal,
   FormattedDate,
   FormattedTime,
@@ -14,10 +19,15 @@ import {
 
 import CheckoutNoteModal from './components/CheckoutNoteModal';
 import MultipieceModal from './components/MultipieceModal';
-import { statuses } from './constants';
+import {
+  statuses,
+  statusMessages,
+} from './constants';
+import { shouldStatusModalBeShown } from './util';
 
 class ModalManager extends React.Component {
   static propTypes = {
+    intl: PropTypes.object,
     checkedoutItem: PropTypes.object.isRequired,
     checkoutNotesMode: PropTypes.bool,
     onDone: PropTypes.func.isRequired,
@@ -39,7 +49,7 @@ class ModalManager extends React.Component {
     this.state = { checkedoutItem, checkoutNotesMode };
     this.steps = [
       {
-        validate: this.shouldStatusModalBeShown,
+        validate: () => shouldStatusModalBeShown(this.state.checkedoutItem),
         exec: () => this.setState({ showStatusModal: true }),
       },
       {
@@ -79,16 +89,6 @@ class ModalManager extends React.Component {
     }
 
     return this.props.onDone();
-  }
-
-  shouldStatusModalBeShown = () => {
-    const { checkedoutItem } = this.state;
-
-    return [
-      statuses.LOST_AND_PAID,
-      statuses.MISSING,
-      statuses.WITHDRAWN,
-    ].includes(checkedoutItem?.status?.name);
   }
 
   shouldCheckoutNoteModalBeShown = () => {
@@ -247,46 +247,43 @@ class ModalManager extends React.Component {
   // This modal handles confirmation (check out? Don't check out?)
   // for various special item statuses, e.g. Missing, Withdrawn.
   renderConfirmStatusModal() {
+    const { intl: { formatMessage } } = this.props;
     const {
       checkedoutItem,
       showStatusModal,
     } = this.state;
+
     const {
       barcode,
       title,
       materialType,
+      status: { name },
     } = checkedoutItem;
+
+    const status = formatMessage({ id: statusMessages[name] });
+
     const values = {
       title,
       barcode,
       materialType: upperFirst(materialType?.name ?? ''),
+      status,
     };
     const messageId = checkedoutItem.discoverySuppress ?
       'ui-checkout.confirmStatusModal.suppressedMessage' :
       'ui-checkout.confirmStatusModal.notSuppressedMessage';
-    values.status = checkedoutItem?.status?.name;
 
-    let heading;
-    switch (checkedoutItem?.status?.name) {
-      case statuses.LOST_AND_PAID:
-        heading = 'ui-checkout.confirmStatusModal.heading.lostAndPaid';
-        break;
-      case statuses.MISSING:
-        heading = 'ui-checkout.confirmStatusModal.heading.missing';
-        break;
-      case statuses.WITHDRAWN:
-        heading = 'ui-checkout.confirmStatusModal.heading.withdrawn';
-        break;
-      default:
-        break;
-    }
+    const heading = (
+      <FormattedMessage
+        id="ui-checkout.confirmStatusModal.heading"
+        values={{ status: lowerFirst(status) }}
+      />);
 
     return (
       <ConfirmationModal
         id="test-confirm-status-modal"
         open={showStatusModal}
         item={checkedoutItem}
-        heading={<FormattedMessage id={heading} />}
+        heading={heading}
         message={<SafeHTMLMessage
           id={messageId}
           values={values}