+++ src/ScanItems.js
@@ -3,7 +3,10 @@ import moment from 'moment';
 import PropTypes from 'prop-types';
 import ReactAudioPlayer from 'react-audio-player';
 import { FormattedMessage } from 'react-intl';
-import { get, isEmpty } from 'lodash';
+import {
+  get,
+  isEmpty,
+} from 'lodash';
 
 import { Icon } from '@folio/stripes/components';
 import { escapeCqlValue } from '@folio/stripes/util';
@@ -104,7 +107,7 @@ class ScanItems extends React.Component {
       loading: false,
       checkoutStatus: null,
       item: null,
-      error: null,
+      errors: [],
     };
   }
 
@@ -133,41 +136,47 @@ class ScanItems extends React.Component {
       openBlockedModal
     } = this.props;
 
+    const errors = [];
+
     if (!barcode) {
-      return {
+      errors.push({
         item: {
           barcode: <FormattedMessage id="ui-checkout.missingDataError" />
         }
-      };
+      });
     }
 
     if (!patron) {
       this.triggerPatronFormSubmit();
-      return {
+      errors.push({
         patron: {
           identifier: <FormattedMessage id="ui-checkout.missingDataError" />
         }
-      };
+      });
     }
 
     if (patronBlocks.length > 0) {
       openBlockedModal();
-      return {
+      errors.push({
         patron: {
           blocked: <FormattedMessage id="ui-checkout.blockModal" />
         }
-      };
+      });
     }
 
-    return {};
+    return errors;
   }
 
   tryCheckout = async (data) => {
     const barcode = get(data, 'item.barcode');
-    const error = this.validate(barcode);
+    const errors = this.validate(barcode);
 
-    if (!isEmpty(error)) {
-      this.setState({ error });
+    if (!isEmpty(this.props.patronBlocks)) {
+      return;
+    }
+
+    if (!isEmpty(errors)) {
+      this.setState({ errors });
       return;
     }
 
@@ -223,19 +232,27 @@ class ScanItems extends React.Component {
   }
 
   override = (data) => {
-    const { mutator: { overrideCheckout } } = this.props;
+    const { mutator: { checkout } } = this.props;
     const { barcode, comment, dueDate } = data;
-    const overrideData = {
-      ...this.getRequestData(barcode),
-      comment,
-      dueDate,
-    };
+    const overrideData = { ...this.getRequestData(barcode) };
 
-    return this.performAction(overrideCheckout, overrideData);
+    if (!dueDate) {
+      overrideData.overrideBlocks = {
+        itemLimitBlock: {},
+        comment,
+      };
+    } else {
+      overrideData.overrideBlocks = {
+        itemNotLoanableBlock: { dueDate },
+        comment,
+      };
+    }
+
+    return this.performAction(checkout, overrideData);
   }
 
   performAction(action, data) {
-    this.setState({ loading: true, error: null });
+    this.setState({ loading: true, errors: [] });
     return action.POST(data)
       .then(this.fetchLoanPolicy)
       .then(this.addScannedItem)
@@ -254,36 +271,11 @@ class ScanItems extends React.Component {
     }
   }
 
-  handleErrors = ({
-    errors: [
-      {
-        parameters,
-        message,
-      } = {},
-    ] = [],
-  }) => {
+  handleErrors = ({ errors }) => {
     // TODO make error message internationalized
     // (https://github.com/folio-org/ui-checkout/pull/408#pullrequestreview-317759489)
-    let itemError;
-    if (!parameters) {
-      itemError = {
-        barcode: <FormattedMessage id="ui-checkout.unknownError" />,
-        _error: 'unknownError',
-      };
-    } else if (parameters.length === 0) {
-      itemError = {
-        barcode: message,
-        loanPolicy: ''
-      };
-    } else {
-      itemError = {
-        barcode: message,
-        _error: parameters[0].key,
-        loanPolicy: parameters[0].value,
-      };
-    }
 
-    this.setState({ error: { item: itemError } });
+    this.setState({ errors });
   }
 
   addScannedItem = (loan) => {
@@ -323,7 +315,7 @@ class ScanItems extends React.Component {
   }
 
   onClearCheckoutErrors = () => {
-    this.setState({ error: null });
+    this.setState({ errors: [] });
   }
 
   onCancel = () => {
@@ -358,7 +350,8 @@ class ScanItems extends React.Component {
       checkoutStatus,
       loading,
       item,
-      checkoutNotesMode
+      errors,
+      checkoutNotesMode,
     } = this.state;
 
     const scannedItems = parentResources.scannedItems || [];
@@ -386,7 +379,7 @@ class ScanItems extends React.Component {
           onSessionEnd={onSessionEnd}
           item={item}
           shouldSubmitAutomatically={shouldSubmitAutomatically}
-          checkoutError={this.state.error}
+          checkoutError={errors}
           onClearCheckoutErrors={this.onClearCheckoutErrors}
           initialValues={initialValues}
         />