+++ src/components/ItemForm/ItemForm.js
@@ -5,14 +5,12 @@ import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
 
 import stripesFinalForm from '@folio/stripes/final-form';
-
 import {
   Col,
   Button,
   Row,
   TextField,
 } from '@folio/stripes/components';
-
 import {
   withStripes,
   stripesShape,
@@ -32,18 +30,19 @@ class ItemForm extends React.Component {
     onOverride: PropTypes.func.isRequired,
     form: PropTypes.object.isRequired,
     formRef: PropTypes.object.isRequired,
-    checkoutError: PropTypes.object,
+    checkoutError: PropTypes.arrayOf(
+      PropTypes.object
+    ).isRequired,
     onClearCheckoutErrors: PropTypes.func,
   };
 
   static defaultProps = {
     patron: {},
     item: {},
-    checkoutError: {},
   };
 
-  static getDerivedStateFromProps(props, { error }) {
-    return { error: props.checkoutError?.item || error || {} };
+  static getDerivedStateFromProps(props, { errors }) {
+    return { errors: props.checkoutError || errors || [] };
   }
 
   constructor(props) {
@@ -52,7 +51,8 @@ class ItemForm extends React.Component {
     this.barcodeEl = React.createRef();
     this.state = {
       overrideModalOpen: false,
-      error: {},
+      errors: [],
+      message: '',
     };
   }
 
@@ -84,12 +84,16 @@ class ItemForm extends React.Component {
     }
   }
 
-  setError = (error) => {
-    this.setState({ error });
+  setError = (errors) => {
+    this.setState({ errors });
   };
 
   openOverrideModal = () => {
-    this.setState({ overrideModalOpen: true });
+    const { errors } = this.state;
+    this.setState({
+      message: errors[0].message,
+      overrideModalOpen: true
+    });
   };
 
   closeOverrideModal = () => {
@@ -106,17 +110,17 @@ class ItemForm extends React.Component {
       onClearCheckoutErrors,
     } = this.props;
 
-    this.setError({});
+    this.setError([]);
     onClearCheckoutErrors();
     form.reset();
   };
 
   onSubmit = async (event) => {
     const { handleSubmit } = this.props;
-    const error = await handleSubmit(event);
+    const errors = await handleSubmit(event);
 
-    if (error?.item) {
-      return error;
+    if (!isEmpty(errors)) {
+      return errors;
     }
 
     return this.clearForm();
@@ -130,7 +134,11 @@ class ItemForm extends React.Component {
       onOverride,
     } = this.props;
 
-    const { error, overrideModalOpen } = this.state;
+    const {
+      errors,
+      message,
+      overrideModalOpen,
+    } = this.state;
     const validationEnabled = false;
 
     return (
@@ -172,25 +180,22 @@ class ItemForm extends React.Component {
             </Col>
           </Row>
         </form>
-        {
-          !isEmpty(error) &&
+        {!isEmpty(errors) &&
           <ErrorModal
             stripes={stripes}
             item={item || {}}
-            message={error.barcode}
-            loanPolicy={error.loanPolicy}
-            open={!isEmpty(error)}
+            errors={errors}
+            open
             openOverrideModal={this.openOverrideModal}
             onClose={this.clearForm}
           />
         }
-        {
-          overrideModalOpen &&
+        {overrideModalOpen &&
           <OverrideModal
             item={item}
+            message={message}
             stripes={stripes}
             onOverride={onOverride}
-            overrideModalOpen={overrideModalOpen}
             closeOverrideModal={this.closeOverrideModal}
           />
         }